# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HiriBot HR API",
        "description": "HiriBot İK (HRIS) platformunun REST API'si. Çalışan yönetimi, izin & devamsızlık, bordro, işe alım, toplantı planlama/kayıt ve denetim logları gibi modüller için güvenli entegrasyon sağlar.",
        "contact": {
            "name": "HiriBot API Destek",
            "email": "support@hiri.ai"
        },
        "license": {
            "name": "HiriBot Proprietary License",
            "url": "https://hiri.ai/legal/api-license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://meetapi.test/api",
            "description": "Prod API Sunucusu"
        },
        {
            "url": "https://staging.hiribot.hiri.ai/api",
            "description": "Staging API Sunucusu"
        },
        {
            "url": "http://hiribot.local/api",
            "description": "Yerel Geliştirme Sunucusu"
        },
        {
            "url": "https://prodapiv2.cmapps.eu"
        }
    ],
    "paths": {
        "/v1/meetings": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Toplantıları listele (sayfalı)",
                "description": "Meeting kayıtlarını sayfalı olarak döner.",
                "operationId": "meetingsIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Sayfa numarası",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Sayfa başı kayıt",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeetingPage"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Toplantı oluştur",
                "operationId": "meetingsStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeetingInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Meeting"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama Hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/v1/meetings/{id}": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Tek toplantıyı getir",
                "operationId": "meetingsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Meeting UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Meeting"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Toplantıyı güncelle",
                "operationId": "meetingsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Meeting UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeetingInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Meeting"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama Hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Toplantıyı sil",
                "description": "Soft delete uygular.",
                "operationId": "meetingsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Meeting UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Silindi."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Meeting": {
                "required": [
                    "id",
                    "link",
                    "type",
                    "start_at",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "7d3c1c7e-02aa-4a4d-bb13-6a6c1a3f4c1a"
                    },
                    "link": {
                        "type": "string",
                        "format": "uri",
                        "maxLength": 1024,
                        "example": "https://meet.google.com/abc-defg-hij"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "meet",
                            "teams",
                            "zoom"
                        ],
                        "example": "meet"
                    },
                    "start_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-01T12:00:00Z"
                    },
                    "end_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-01T13:00:00Z",
                        "nullable": true
                    },
                    "record_path": {
                        "type": "string",
                        "maxLength": 2048,
                        "example": "/records/2025/09/01/abc123.mp4",
                        "nullable": true
                    },
                    "log": {
                        "type": "object",
                        "example": {
                            "scheduled_by": "a8c2...",
                            "status": "queued"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MeetingInput": {
                "required": [
                    "link",
                    "type",
                    "start_at"
                ],
                "properties": {
                    "link": {
                        "type": "string",
                        "format": "uri",
                        "maxLength": 1024,
                        "example": "https://meet.google.com/abc-defg-hij"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "meet",
                            "teams",
                            "zoom"
                        ],
                        "example": "teams"
                    },
                    "start_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-01T12:00:00Z"
                    },
                    "end_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-01T13:30:00Z",
                        "nullable": true
                    },
                    "record_path": {
                        "type": "string",
                        "maxLength": 2048,
                        "example": "https://cdn.hiribot.hiri.ai/recs/abc.mp4",
                        "nullable": true
                    },
                    "log": {
                        "type": "object",
                        "example": {
                            "agent": "recorder-1",
                            "priority": 5
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "link": [
                                "The link field must be a valid URL."
                            ]
                        },
                        "additionalProperties": {
                            "schema": "MeetingController",
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "MeetingPage": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Meeting"
                        }
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token ile yetkilendirme. Örn: 'Authorization: Bearer {token}'",
                "bearerFormat": "Sanctum/JWT",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "Proje düzeyi API anahtarı. Tüm isteklerde 'X-API-KEY' başlığı gerektirir.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}